Fix clang-tidy performance-enum-size warnings in headers

+++ src/List.h
@@ -30,7 +30,7 @@
 
 namespace zeek {
 
-enum class ListOrder : int { ORDERED, UNORDERED };
+enum class ListOrder : uint8_t { ORDERED, UNORDERED };
 
 template<typename T, ListOrder Order = ListOrder::ORDERED>
 class List {
